Move codes to pluginUpdater.cpp and pluginUpdater.h.
